// src/app/models/generator-anually.ts
export class GeneratorAnually {
    key: string;
    year: string;
    machine1ControlUnitCaseCleaning: string;
    machine1InsulationResistanceVerifi: string;
    machine1CheckTightnessofallPower: string;
    machine1CheckingofExcisterAirGap: string;
    machine2ControlUnitCaseCleaning: string;
    machine2InsulationResistanceVerifi: string;
    machine2CheckTightnessofallPower: string;
    machine2CheckingofExcisterAirGap: string;

}

export class Machine {

}